+++ Source/Core/HeadingPitchRange.js
@@ -1,10 +1,10 @@
 /*global define*/
 define([
-        '../Core/HeadingPitchRange',
-        '../Core/deprecationWarning'
+        './defaultValue',
+        './defined'
     ], function(
-        CoreHeadingPitchRange,
-        deprecationWarning) {
+        defaultValue,
+        defined) {
     "use strict";
 
     /**
@@ -18,12 +18,26 @@ define([
      * @param {Number} [heading=0.0] The heading angle in radians.
      * @param {Number} [pitch=0.0] The pitch angle in radians.
      * @param {Number} [range=0.0] The distance from the center in meters.
-     *
-     * @deprecated
      */
     var HeadingPitchRange = function(heading, pitch, range) {
-        deprecationWarning('HeadingPitchRange', 'Scene/HeadingPitchRange is deprecated. Use Core/HeadingPitchRange instead.');
-        return new CoreHeadingPitchRange(heading, pitch, range);
+        /**
+         * Heading is the rotation from the local north direction where a positive angle is increasing eastward.
+         * @type {Number}
+         */
+        this.heading = defaultValue(heading, 0.0);
+
+        /**
+         * Pitch is the rotation from the local xy-plane. Positive pitch angles
+         * are above the plane. Negative pitch angles are below the plane.
+         * @type {Number}
+         */
+        this.pitch = defaultValue(pitch, 0.0);
+
+        /**
+         * Range is the distance from the center of the local frame.
+         * @type {Number}
+         */
+        this.range = defaultValue(range, 0.0);
     };
 
     /**
@@ -34,7 +48,17 @@ define([
      * @returns {HeadingPitchRange} The modified result parameter or a new HeadingPitchRange instance if one was not provided. (Returns undefined if hpr is undefined)
      */
     HeadingPitchRange.clone = function(hpr, result) {
-        return CoreHeadingPitchRange.clone(hpr, result);
+        if (!defined(hpr)) {
+            return undefined;
+        }
+        if (!defined(result)) {
+            result = new HeadingPitchRange();
+        }
+
+        result.heading = hpr.heading;
+        result.pitch = hpr.pitch;
+        result.range = hpr.range;
+        return result;
     };
 
     return HeadingPitchRange;